fix: Support for BMFF V3 hashing options (#1253)
* Support for BMFF Hash V3
validation of fixedBlockSize and fixedBlockSizes Merkle trees
generation of fixedBlockSize Merkle trees

* Bug fix for case when BMFF exclusion subset is using length with no offset

* Setting to control Merkle tree storage

* Change error message

* More error checking
PR comment change

* Make settings fetch return error is not set

* Add test for no fixed offset

* Fix formatting

* Check for conflicting block types

* Fix error response for competing Merkle block types

* Make spec responses

* Allow 0 length fixed offset

* A little cleanup

* Make sure there is a recognized hash binding

* Fix for some more edge cases

* clippy fix

---------

+++ sdk/src/assertions/box_hash.rs
@@ -40,10 +40,10 @@ pub struct BoxMap {
     pub pad: ByteBuf,
 
     #[serde(skip)]
-    pub range_start: usize,
+    pub range_start: u64,
 
     #[serde(skip)]
-    pub range_len: usize,
+    pub range_len: u64,
 }
 
 /// Helper class to create BoxHash assertion
@@ -106,7 +106,7 @@ impl BoxHash {
 
             // build up current inclusion, consuming all names in this BoxMap
             let mut skip_c2pa = false;
-            let mut inclusion = HashRange::new(0, 0);
+            let mut inclusion = HashRange::new(0u64, 0u64);
             for name in &bm.names {
                 match source_bms.get(source_index) {
                     Some(next_source_bm) => {
